requirements/base.txt
django==4.2.11
psycopg2-binary
python-environ
django-cors-headers
django-jazzmin
drf-yasg
djangorestframework
django-filter
Pillow
djangorestframework-simplejwt
channels
channels_redis
django-ckeditor
daphne
django-debug-toolbar


